app/requirements.txt
psycopg2-binary>=2.9.5
ttkbootstrap